% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy}
\alias{accuracy}
\title{Accuracy}
\usage{
accuracy(x, propvar, propopt)
}
\arguments{
\item{x}{A dataset}

\item{propvar}{The proportion variable as character vector}

\item{propopt}{The proportion option as character vector}
}
\description{
Function to calculate the accuracy of a sample. The function returns an e value
}

