% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basedata.R
\name{assemble.basedata}
\alias{assemble.basedata}
\title{Assemble Basedata}
\usage{
assemble.basedata(df = haq, select.idx = which(haq$hh.BM.use == "BM"),
  beforevar = "coal.units.winter.before.bm.0",
  aftervar = "coal.units.winter.after.bm.0",
  currentvar = "current.coal.units.winter.0", vars = c("response.id",
  "town", "surveyyear", "bm.usage.start.year.0", "bmstartyearmonth",
  "coal.buying.format.r", "coal.units.winter.before.bm.0",
  "coal.units.winter.after.bm.0", "current.coal.units.winter.0"),
  shortnames = c("response.id", "town", "surveyyear", "BM start year",
  "BM start date", "Coal format", "winter before BM", "winter after BM",
  "Current winter", "eef"), verbose = FALSE)
}
\arguments{
\item{df}{Data frame containing HAQ survey data}

\item{select.idx}{The selection indices for people who use BM as a numeric}

\item{beforevar}{The variable referring to coal use before BM as character vector}

\item{aftervar}{The variable referring to coal use after BM as character vector}

\item{currentvar}{The variable referring to current coal use as character vector}

\item{vars}{The variables under consideration for creating the new data frame for only
the selection indices as a character vector}

\item{shortnames}{Shorter names to replace the longer variable names}

\item{verbose}{Displays function messages when TRUE}
}
\description{
Assemble the data needed to make the baseline calculation
}

