% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basedata.R
\name{assemble.qcdata.2}
\alias{assemble.qcdata.2}
\title{Assemble QC Data 2}
\usage{
assemble.qcdata.2(df = haq, select.idx = which(df[, "hh.BM.use"] == "BM"),
  vars = c("response.id", "contact.number.0", "respondent.name.0",
  "hh.surname.0", "town", "fieldworker", "surveyyear", "hh.coal.use.r",
  "hh.BM.use", "bm.saved.coal.winter.r", "bmstartyearmonth", "bmstartyearmonth",
  "coal.buying.format.r", "coal.units.winter.before.bm.0",
  "coal.units.winter.after.bm.0", "current.coal.units.winter.0"),
  shortnames = c("response.id", "Contact number", "Name", "Surname", "town",
  "Fieldworker", "surveyyear", "Coal use", "BM use", "BM saved coal winter",
  "BM start year", "BM start date", "Coal format", "winter before BM",
  "winter after BM", "Current winter", "eef"))
}
\arguments{
\item{df}{Data frame containing HAQ survey data}

\item{select.idx}{The selection indices for people who use BM as a numeric}

\item{vars}{The variables under consideration for creating a new data frame that
will only contain their values for each selection index}

\item{shortnames}{Shorter names to replace the longer variable names}
}
\description{
Assemble the data needed to make the baseline calculation
}

