% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.summary.R
\name{att.summary}
\alias{att.summary}
\title{Attendance Summary}
\usage{
att.summary(df, var = "bm.demo.attendance.r", value = "YES", pop = pop,
  group = haq$town)
}
\arguments{
\item{df}{Data frame containing BM data}

\item{var}{The variable corresponding to BM demo attendance}

\item{value}{The value of concern for the supplied var}

\item{pop}{Data frame containing population}

\item{group}{The grouping variable}
}
\description{
Function to summarise attendance
}

