% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleannr.R
\name{base.number}
\alias{base.number}
\title{Base Number}
\usage{
base.number(x)
}
\arguments{
\item{x}{A character vector where matches are sought, or an
object which can be coerced by as.character to a character vector. Long vectors are supported}
}
\description{
Find original stand of subdivided stands write with / or .
}

