% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Baseline.tables.R
\name{baseline.sample}
\alias{baseline.sample}
\title{Baseline Sample}
\usage{
baseline.sample(haq)
}
\arguments{
\item{haq}{Data frame containing HAQ survey data}
}
\description{
Creates sample tables for inspection
}

