% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Baseline.tables.R
\name{baseline.tables}
\alias{baseline.tables}
\title{Baseline Tables}
\usage{
baseline.tables(haq, basedata, uta, cl.alt = get("cl.alt", pos =
  sys.frame(which = 1)), cl.coal = get("cl.coal", pos = sys.frame(which = 1)),
  Uother, basedir, tabdir, verbose = TRUE)
}
\arguments{
\item{haq}{Data frame containing HAQ survey data}

\item{basedata}{Data frame containing baseline data}

\item{uta}{Data frame containing UTA data}

\item{cl.alt}{Data frame containing 90 percent confidence
interval data for the estimation of BM users attributable to the project
activity in the Target Area}

\item{cl.coal}{Data frame containing 90 percent confidence interval data
for number of coal users in the Target Area}

\item{Uother}{Data frame of data describing coal users converted by other sources}

\item{basedir}{The baseline directory as character vector}

\item{tabdir}{Directory to which tables are saved within the baseline directory
as character vector}

\item{verbose}{Displays function messages when TRUE}
}
\description{
Function creates tables for report
}

