% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.vars.R
\name{baseline.vars}
\alias{baseline.vars}
\title{Baseline Variables}
\usage{
baseline.vars(haq = get("haq", envir = sys.frame(which = 0)),
  basedata = get("basedata", envir = sys.frame(which = 0)),
  cl.alt = get("cl.alt", pos = sys.frame(which = 0)),
  basedir = get("basedir", envir = sys.frame(which = 0)),
  tabdir = get("tabdir", sys.frame(which = 0)), meth = ifelse(exists("meth",
  where = sys.frame(which = -2)), get("meth", sys.frame(which = -2)), c("old",
  "new")[1]))
}
\arguments{
\item{haq}{Data frame containing HAQ survey data}

\item{basedata}{Data frame containing baseline data}

\item{cl.alt}{Data frame containing 90 percent confidence
interval data for the estimation of BM users attributable to the project
activity in the Target Area}

\item{basedir}{The baseline directory as character vector}

\item{tabdir}{Directory to which tables are saved withing the baseline directory
as character vector}

\item{meth}{The method to be used, "old" or "new"}
}
\description{
Creates summary tables for variables related to BM use
}
\note{
By default this function will retrieve all of its information from the global environment
}

