% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTA.R
\name{cl.Ualt}
\alias{cl.Ualt}
\title{CI Ualt}
\usage{
cl.Ualt(df, Coal = "hh.coal.use.r", coalvar = "YES", var = "hh.BM.use",
  pop = pop, group = haq$town, othervar = "introduction.to.bm.r",
  altvar = "hh.BM.use")
}
\arguments{
\item{df}{Data frame containing survey data}

\item{Coal}{The variable associated with whether a household uses coal}

\item{coalvar}{The value of the coal use variable of importance}

\item{var}{The variable associated with whether a household uses coal}

\item{pop}{Data frame containing population data}

\item{group}{The grouping variable}

\item{othervar}{Any other variable to be considered}

\item{altvar}{The alternative variable that could also be present}
}
\description{
Creates 90 percent confidence interval for Ualt
}

