% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTA.R
\name{cl.UTA}
\alias{cl.UTA}
\title{CI UTA}
\usage{
cl.UTA(df, Coal = "hh.coal.use.r", var = "hh.coal.use.r", coalvar = "Yes",
  pop = pop, group = haq$town, othervar = "introduction.to.bm")
}
\arguments{
\item{df}{Data frame containing survey data}

\item{Coal}{The variable associated with whether a household uses coal}

\item{var}{The variable associated with whether a household uses coal}

\item{coalvar}{The value of the coal use variable of importance}

\item{pop}{Data frame containing population data}

\item{group}{The grouping variable}

\item{othervar}{Any other variable to be considered}
}
\description{
Creates 90 percent confidence interval for UTA
}

