% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleannr.R
\name{cleannr}
\alias{cleannr}
\title{Cleaner}
\usage{
cleannr(df, mainplace_name = "mainplace_name", standvar = "standnumber",
  surnamevar = "surname")
}
\arguments{
\item{df}{Data frame containing survey data}

\item{mainplace_name}{The variable containing the main place names as character vector}

\item{standvar}{The variable containing stand numbers as character vector}

\item{surnamevar}{The variable containing surnames as character vector}
}
\description{
Function creates a vector of clean (i.e. remove letter at end) stand numbers
}
\examples{
ermelo=cleannr(ermelo,"mainplace_name","standnumber","surname")
}

