% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.summary.R
\name{coal.summary.cl}
\alias{coal.summary.cl}
\title{Coal Summary CI}
\usage{
coal.summary.cl(df, BM = "hh.BM.use", var = "total.winter.save",
  con = "BM", coef = 2.2099, pop = pop, group = haq$town)
}
\arguments{
\item{df}{Data frame containing BM data}

\item{BM}{The BM use variable as character vector}

\item{var}{The variable corresponding to seasonal savings}

\item{con}{The value that corresponds to active BM use}

\item{coef}{The coefficient used to calculate variable CO2 emissions}

\item{pop}{Data frame containing population data}

\item{group}{The grouping variable}
}
\description{
Creates coal use summary with Confidence Interval included
}

