% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er.disc.table.R
\name{er.disc.table}
\alias{er.disc.table}
\title{ER Discount Table}
\usage{
er.disc.table(projectname = get("projectname", envir = sys.frame(which = 0)),
  basedir = get("basedir", envir = sys.frame(which = 0)),
  tabdir = get("tabdir", envir = sys.frame(which = 0)))
}
\arguments{
\item{projectname}{name of file to create without the .tex
extension. If this option is not set, value/string of x (see above) is
printed in the top left corner of the table. Set title=&rdquo; to suppress this output. By
default the name of the project will be retrieved from the global environment}

\item{basedir}{The baseline directory as character vector}

\item{tabdir}{Directory to which tables are saved within the baseline directory
as character vector}
}
\description{
Crates a table of calender year adjusted emission reductions in CO2
}

