% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er.disc.table.R
\name{er.jan.april}
\alias{er.jan.april}
\title{ER Jan April}
\usage{
er.jan.april(df = haq, bmvar = "hh.summer.BM.use",
  yearvar = "year.of.initiation.0",
  summerconbaseyear = "summerconbase.year",
  summercurrentyear = "summer.current.year", startyear = 2010,
  basedirr = basedir, tabdirr = tabdir, year = 2011)
}
\arguments{
\item{df}{Data frame containing HAQ survey data}

\item{bmvar}{The Basa Magogo household summer use variable as character vector}

\item{yearvar}{The variable referring to the year of initiation as character vector}

\item{summerconbaseyear}{The base summer year variable as character vector}

\item{summercurrentyear}{The current summer year variable as character vector}

\item{startyear}{The start year as a numeric}

\item{basedirr}{The baseline directory as character vector}

\item{tabdirr}{Directory to which tables are saved within the baseline directory
as character vector}

\item{year}{The year to appear in the caption of the emissions reduction output table as a numeric}
}
\description{
Function to discount last 4 summer months
}

