% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuburbFind.R
\name{ErfFindSave}
\alias{ErfFindSave}
\title{Erf Find Save}
\usage{
ErfFindSave(SPDF, SuburbBoundaries, var = "MP_NAME",
  outdir = "~/Dropbox/Verifikasie2012/Phumula/2.SampleFrame/")
}
\arguments{
\item{SPDF}{A spacialPolygonsDataFrame that has the boundaries (the polygons) and their names (data)}

\item{SuburbBoundaries}{The boundaries related to the suburb under consideration}

\item{var}{The variable under consideration as a character vector}

\item{outdir}{The desired output directory as character vector}
}
\description{
Function to find the correct suburb of a known address and save it to the chosen directory
}

