% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleannr.R
\name{isolate.number}
\alias{isolate.number}
\title{Isolate Number}
\usage{
isolate.number(x)
}
\arguments{
\item{x}{A character vector where matches are sought, or an
object which can be coerced by as.character to a character vector. Long vectors are supported}
}
\description{
Isolate only the stand number
}

