% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meann.R
\name{meann}
\alias{meann}
\title{Meann}
\usage{
meann(x, ...)
}
\arguments{
\item{x}{The object to have its mean computed}

\item{...}{Further arguments passed to or from other methods}
}
\description{
A simple mean function
}

