% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noplots.R
\name{noplots}
\alias{noplots}
\title{No Plots}
\usage{
noplots(x, standnumbervar = "standnumber", suburbvar = "suburb", SPx,
  TEXT = "TEXT")
}
\arguments{
\item{x}{Data frame containing the stand numbers}

\item{standnumbervar}{The name of the variable that contains stand numbers as character vector}

\item{suburbvar}{The name of the variable that contains suburbs as character vector}

\item{SPx}{A SpatialPolygonsDataFrame containing geographic data}

\item{TEXT}{The name of the text variable as character vector}
}
\description{
Creates varible called noplots that provides suburbs that are not plotted
}

