% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr.plot.report.R
\name{nr.plot.report}
\alias{nr.plot.report}
\title{Stand Number Plot QC Function}
\usage{
nr.plot.report(x, standnumbervar = "standnumber", suburbvar = "suburb", SPx,
  TEXT = "TEXT")
}
\arguments{
\item{x}{Data frame containing the stand numbers}

\item{standnumbervar}{The name of the variable that corresponds to the stand number as character vector}

\item{suburbvar}{The name of the variable that corresponds to the suburb as character vector}

\item{SPx}{A SpatialPolygonsDataFrame containing geographic data}

\item{TEXT}{The name of the text variable as character vector}
}
\description{
Match a list of stand numbers to the stand numbers in a geografic
file (like a SpatialPolygonsDataFrame) and prints a summary of whether
a suburb is plotted or not
}

