% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.formula.table.R
\name{project.formula.table}
\alias{project.formula.table}
\title{Project Formula Table}
\usage{
project.formula.table()
}
\description{
Function to make project formula table and ER table for report. This function retrieves the needed
data from the global environment such as the base and table directories and the HAQ data frame.
}

