% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.sample.R
\name{project.sample}
\alias{project.sample}
\title{Project Sample}
\usage{
project.sample()
}
\description{
Creates a sample of relevant variables for inspection as well as a summary of project
emission variables
}
\note{
The function contains no arguments and relies on the haq data frame as well as
base and table directories being present in the global environmet in order to work.
}

