% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.summary.R
\name{project.summary.09.proj}
\alias{project.summary.09.proj}
\title{Project Summary 09 Projection}
\usage{
project.summary.09.proj(df, BM = "hh.BM.use", var = "total.winter.save",
  coef = 2.2099, pop = pop, group = haq$town, datevar = "received")
}
\arguments{
\item{df}{Data frame containing BM data}

\item{BM}{The BM use variable as character vector}

\item{var}{The variable corresponding to seasonal savings}

\item{coef}{The coefficient used to calculate variable CO2 emissions}

\item{pop}{Data frame containing population data}

\item{group}{The grouping variable}

\item{datevar}{The variable corresponding to the dates under consideration}
}
\description{
Function to summarise BM use, coal use and Emission results
including projected discounts and percentages CO2
}

