% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTA.R
\name{project.summary.UTA}
\alias{project.summary.UTA}
\title{Project summary UTA}
\usage{
project.summary.UTA(df, Coal = "hh.coal.use.r", coalvar = "YES",
  var = "hh.coal.use.r", pop = pop, group = haq$town,
  othervar = "introduction.to.bm.r", altvar = "hh.BM.use", coal = TRUE,
  BM = TRUE, other = TRUE)
}
\arguments{
\item{df}{Data frame containing survey data}

\item{Coal}{The variable associated with whether a household uses coal}

\item{coalvar}{The value of the coal use variable of importance}

\item{var}{The variable associated with whether a household uses coal}

\item{pop}{Data frame containing population data}

\item{group}{The grouping variable}

\item{othervar}{Any other variable to be considered}

\item{altvar}{The alternative variable that could also be present}

\item{coal}{If FALSE the function omits coal users from the table}

\item{BM}{If If FALSE the function omits BM users from the table}

\item{other}{If FALSE the function omits other users from the table}
}
\description{
Function to calculate UTA
}

