% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.tables.R
\name{project.tables}
\alias{project.tables}
\title{Project Tables}
\usage{
project.tables(haq = get("haq", envir = sys.frame(which = 0)),
  Ualt = get("Ualt", envir = sys.frame(which = 0)))
}
\arguments{
\item{haq}{Data frame containing HAQ survey data}

\item{Ualt}{Data frame containing alternative usage information}
}
\description{
Function creates summary tables such as summer project coal use by town and also
assigns Project coal consumptions and emissions by town for different per season to
the global environment.
}

