% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.var.table.R
\name{project.var.table}
\alias{project.var.table}
\title{Project variable Table}
\usage{
project.var.table()
}
\description{
Function to print summary of project variables and save it to the table directory
as specified in the global environment
}
\note{
The function contains no arguments and relies on the haq data frame as well as
base and table directories being present in the global environmet in order to work.
}

