% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refact.R
\name{refact.df}
\alias{refact.df}
\title{Refactor Data Frame}
\usage{
refact.df(x)
}
\arguments{
\item{x}{Object to have its levlels dropped}
}
\description{
Function drops unneeded factor levels of a data frame
}

