% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.table.R
\name{report.table}
\alias{report.table}
\title{Report Table}
\usage{
report.table(x, caption, label, directory)
}
\arguments{
\item{x}{An R object of class found among methods(xtable).
See below on how to write additional method functions for xtable}

\item{caption}{Character vector of length 1 or 2 containing the table's
caption or title. If length is 2, the second item is the "short caption"
used when LaTeX generates a "List of Tables". Set to NULL to suppress
the caption. Default value is NULL}

\item{label}{Character vector of length 1 containing the LaTeX label or HTML anchor.
Set to NULL to suppress the label. Default value is NULL}

\item{directory}{Name of file where the resulting code should be saved.
If file="", output is displayed on screen. Note that the function
also (invisibly) returns a character vector of the results (which can be helpful for post-processing).
Default value is ""}
}
\description{
Function to export a Latex table with caption and label
}

