% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws.R
\name{ss}
\alias{ss}
\title{Summer Savings}
\usage{
ss(x)
}
\arguments{
\item{x}{An R object (Not used in this function)}
}
\description{
Function to make summer savings table in final report
}
\note{
This function will require the BMbuurt and haq data frames to be present in the global
invironment in order to work
}

