% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat90.R
\name{strat90}
\alias{strat90}
\title{Population variance per Strate with 90 percent CI}
\usage{
strat90(Households, Sample, pointest, debug = TRUE)
}
\arguments{
\item{Households}{Data frame containing household survey data}

\item{Sample}{Data frame containing sample information}

\item{pointest}{Data frame containing point test information}

\item{debug}{Performs the necessary global assignments when TRUE}
}
\description{
Function creates a table of estimated proportion variance per strata, including
intervals and a 90 percent confidence interval
}

