% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuburbFind.R
\name{SuburbFind}
\alias{SuburbFind}
\title{Find Suburb}
\usage{
SuburbFind(SPDF, SuburbBoundaries, var = "MP_NAME")
}
\arguments{
\item{SPDF}{A spacialPolygonsDataFrame that has the boundaries (the polygons) and their names (data)}

\item{SuburbBoundaries}{The boundaries related to the suburb under consideration}

\item{var}{The variable under consideration as a character vector}
}
\description{
Function to find the correct suburb of a known address
}
\examples{
SPDF$Suburnames <- SuburbFind(SPDF,SuburbBoundaries)
}

