% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryBy.table.R
\name{summaryBy.table}
\alias{summaryBy.table}
\title{SummaryBy Table}
\usage{
summaryBy.table(df, varname = "CPy", fn = "yearprojectbytown",
  label = "projectbytownyear",
  caption = "Annual Project coal consumption by town")
}
\arguments{
\item{df}{Data frame to generate summary statistics for}

\item{varname}{The variable name under consideration}

\item{fn}{The desired file name for the output LaTeX table}

\item{label}{Character vector of length 1 containing the
LaTeX label or HTML anchor. Set to NULL to suppress the label. Default value is NULL.}

\item{caption}{Character vector of length 1 or 2 containing the
table's caption or title. If length is 2, the second item is the "short caption"
used when LaTeX generates a "List of Tables". Set to NULL to suppress the caption.
Default value is NULL.}
}
\description{
Function to calculate groupwise summary statistics
}

