% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summry.R
\name{summry}
\alias{summry}
\title{Summary}
\usage{

  summry(file = "~/Desktop/Nova/DEE/GS/PDD/ROFWERK/bsagld-101214-CJP-PDD-spreistaat-basesummary.csv")
}
\arguments{
\item{file}{The file path to the base summary csv}
}
\description{
Function to summarise ER calculation results
}

