% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meann.R
\name{sumn}
\alias{sumn}
\title{Sumn}
\usage{
sumn(x, ...)
}
\arguments{
\item{x}{The object to have its sum computed}

\item{...}{Further arguments passed to or from other methods}
}
\description{
A simple sum function
}

