% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.date.table.R
\name{survey.date.table}
\alias{survey.date.table}
\title{Survey Date Table}
\usage{
survey.date.table(haq = get("haq", envir = environment(survey.date.table)),
  basedir = get("basedir", envir = environment(survey.date.table)),
  tabdir = get("tabdir", envir = environment(survey.date.table)))
}
\arguments{
\item{haq}{Data frame containing HAQ survey data (retreives from global environment
by default)}

\item{basedir}{The base directory where files created are to be saved}

\item{tabdir}{The directory to which latex tables will be saved}
}
\description{
Creates a table that statistically describe the survey dates per town
}

