% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yapply.R
\name{yapply}
\alias{yapply}
\title{Yapply}
\usage{
yapply(X, FUN, ...)
}
\arguments{
\item{X}{an atomic object, typically a vector}

\item{FUN}{Function to apply}

\item{...}{Optional arguments to FUN: the Note section.}
}
\description{
This function was provided by Romain Francois based on something by Thomas Lumley
}
\note{
Optional arguments to FUN supplied by the ... argument
are not divided into cells. It is therefore inappropriate for
FUN to expect additional arguments with the same length as X.
}

